Move a bunch of shared project scope services to a shared provider
